Taxanomic macrophyte data

A dataset containing one macrophyte sample.

taxa_data

Format

A data frame with 10 rows and 11 variables:

location_id

Unique ID of location sampled

sample_id

Unique ID of sample

date_taken

Date the sample was taken

question

What is question is being asked, in this instance `Percentage Cover Band` of a given taxon

response

Value collect in reponse to question

taxon

The taxon associated with the question being asked if applicable

alkalinity

alkalinity value

source_altitude

Source altitude

dist_from_source

Distance from source

slope

Slope

quality_element

Quality element